Change approach to function values
Function value from now on plays strictly a role of a getter.
Function type was removed (although we may bring it back if we feel
that function operations that are not conceptually getters can make
valid representation of database values).

Default namespace in relation is now root, it simplifies validation and normalization.