Always define relations at base level
It makes complex changes to relations possible, e.g. values then can be
restored in any order on every prototype level. Additionally to that
'foreign' tag for relations is provided, it's added to help handlie
relations by some custom mechanisms. First (and might be the last) use
case for that is reverse relations, which now are not blocked by
existing relations of same name.